cleaned up the code by adding a daily cell custom uitableviewcell for section 0 or "Daily"

+++ TVCCYA/Objectives/ObjectivesController+UITableView.swift
@@ -11,40 +11,51 @@ import UIKit
 extension ObjectivesController {
     
     override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        let objective = allObjectives[indexPath.section][indexPath.row]
-        let progressionController = ProgressionController()
-        progressionController.objective = objective
-        navigationController?.pushViewController(progressionController, animated: true)
+        if indexPath.section == 0 {
+
+        } else {
+            let objective = allObjectives[indexPath.section].objectives[indexPath.row]
+            let progressionController = ProgressionController()
+            progressionController.objective = objective
+            navigationController?.pushViewController(progressionController, animated: true)
+        }
     }
     
     override func tableView(_ tableView: UITableView, editActionsForRowAt indexPath: IndexPath) -> [UITableViewRowAction]? {
         
-        let deleteAction = UITableViewRowAction(style: .destructive, title: "Done") { (_, indexPath) in
-            let objective = self.allObjectives[indexPath.section][indexPath.row]
+        var rowActions = [UITableViewRowAction]()
+        
+        if indexPath.section == 0 {
             
-            // remove the task from our tableView
-            self.allObjectives[indexPath.section].remove(at: indexPath.row)
-            self.tableView.deleteRows(at: [indexPath], with: .automatic)
-            
-            // delete the objective from core data
-            let context = CoreDataManager.shared.persistentContainer.viewContext
-            
-            context.delete(objective)
-            
-            do {
-                try context.save()
-            } catch let saveErr {
-                print("Failed to delete objective:", saveErr)
+        } else {
+        
+            let deleteAction = UITableViewRowAction(style: .destructive, title: "Done") { (_, indexPath) in
+                let objective = self.allObjectives[indexPath.section].objectives[indexPath.row]
+                
+                // remove the task from our tableView
+                self.allObjectives[indexPath.section].objectives.remove(at: indexPath.row)
+                self.tableView.deleteRows(at: [indexPath], with: .automatic)
+                
+                // delete the objective from core data
+                let context = CoreDataManager.shared.persistentContainer.viewContext
+                
+                context.delete(objective)
+                
+                do {
+                    try context.save()
+                } catch let saveErr {
+                    print("Failed to delete objective:", saveErr)
+                }
             }
+            
+            deleteAction.backgroundColor = UIColor.doneGreen
+            
+            let editAction = UITableViewRowAction(style: .normal, title: "Edit", handler: editHandlerFunction)
+            
+            editAction.backgroundColor = UIColor.mainBlue
+            rowActions = [deleteAction, editAction]
         }
-        
-        deleteAction.backgroundColor = UIColor.doneGreen
-        
-        let editAction = UITableViewRowAction(style: .normal, title: "Edit", handler: editHandlerFunction)
-        
-        editAction.backgroundColor = UIColor.mainBlue
-        
-        return [deleteAction, editAction]
+        return rowActions
     }
     
     private func editHandlerFunction(action: UITableViewRowAction, indexPath: IndexPath) {
@@ -53,7 +64,7 @@ extension ObjectivesController {
         
         enterObjectiveController.delegate = self
         
-        enterObjectiveController.objective = allObjectives[indexPath.section][indexPath.row]
+        enterObjectiveController.objective = allObjectives[indexPath.section].objectives[indexPath.row]
         
         navigationController?.pushViewController(enterObjectiveController, animated: true)
     }
@@ -66,7 +77,7 @@ extension ObjectivesController {
         label.font = UIFont.boldSystemFont(ofSize: 16)
         return label
     }
-    
+
     override func tableView(_ tableView: UITableView, heightForFooterInSection section: Int) -> CGFloat {
         if section == 0 {
             return objectives.count == 0 ? 75 : 0
@@ -81,10 +92,14 @@ extension ObjectivesController {
         let yearly = objectives.filter { $0.type == "Yearly" }
         
         allObjectives = [
-            daily,
-            weekly,
-            monthly,
-            yearly
+            ExpandableObjectives(isExpanded: true, objectives: daily),
+            ExpandableObjectives(isExpanded: true, objectives: weekly),
+            ExpandableObjectives(isExpanded: true, objectives: monthly),
+            ExpandableObjectives(isExpanded: true, objectives: yearly)
+//            daily,
+//            weekly,
+//            monthly,
+//            yearly
         ]
     }
     
@@ -97,35 +112,42 @@ extension ObjectivesController {
         view.backgroundColor = UIColor.extraLightBlue
         
         let button = UIButton()
+        button.addTarget(self, action: #selector(handleExpandClose), for: .touchUpInside)
+        button.tag = section
         
         let label = UILabel()
         label.font = UIFont.boldSystemFont(ofSize: 20)
         label.textColor = UIColor.mainLightBlue
         
+        let rowCountLabel = UILabel()
+        rowCountLabel.textAlignment = .center
+        
+        let attributedText = NSMutableAttributedString(string: "[", attributes: [NSAttributedStringKey.font: UIFont.boldSystemFont(ofSize: 20), NSAttributedStringKey.foregroundColor: UIColor.mainLightGray])
+        
+        attributedText.append(NSAttributedString(string: "\(allObjectives[section].objectives.count)", attributes: [NSAttributedStringKey.font: UIFont.boldSystemFont(ofSize: 16), NSAttributedStringKey.foregroundColor: UIColor.mainBlue]))
+        
+        attributedText.append(NSAttributedString(string: "]", attributes: [NSAttributedStringKey.font: UIFont.boldSystemFont(ofSize: 20), NSAttributedStringKey.foregroundColor: UIColor.mainLightGray]))
+        
+        rowCountLabel.attributedText = attributedText
+        
         if section == 0 {
             label.text = "Daily"
-//            button.setImage(#imageLiteral(resourceName: "dailyPlus").withRenderingMode(.alwaysOriginal), for: .normal)
-//            button.addTarget(self, action: #selector(handleDailyObjective), for: .touchUpInside)
         } else if section == 1 {
             label.text = "Weekly"
-//            button.setImage(#imageLiteral(resourceName: "weeklyPlus").withRenderingMode(.alwaysOriginal), for: .normal)
-//            button.addTarget(self, action: #selector(handleObjective), for: .touchUpInside)
         } else if section == 2 {
             label.text = "Monthly"
-//            button.setImage(#imageLiteral(resourceName: "monthlyPlus").withRenderingMode(.alwaysOriginal), for: .normal)
-//            button.addTarget(self, action: #selector(handleObjective), for: .touchUpInside)
         } else {
             label.text = "Yearly"
-//            button.setImage(#imageLiteral(resourceName: "yearlyPlus").withRenderingMode(.alwaysOriginal), for: .normal)
-//            button.addTarget(self, action: #selector(handleObjective), for: .touchUpInside)
         }
         
-        view.addSubview(label)
-        label.anchor(top: view.topAnchor, left: view.leftAnchor, bottom: view.bottomAnchor, right: nil, paddingTop: 0, paddingLeft: 16, paddingBottom: 0, paddingRight: 0, width: 0, height: 0)
-        
         view.addSubview(button)
-        button.anchor(top: nil, left: nil, bottom: nil, right: view.rightAnchor, paddingTop: 0, paddingLeft: 0, paddingBottom: 0, paddingRight: 8, width: 40, height: 40)
-        button.centerYAnchor.constraint(equalTo: view.centerYAnchor).isActive = true
+        button.anchor(top: view.topAnchor, left: view.leftAnchor, bottom: view.bottomAnchor, right: view.rightAnchor, paddingTop: 0, paddingLeft: 0, paddingBottom: 0, paddingRight: 0, width: 0, height: 0)
+        
+        button.addSubview(label)
+        label.anchor(top: button.topAnchor, left: button.leftAnchor, bottom: button.bottomAnchor, right: nil, paddingTop: 0, paddingLeft: 16, paddingBottom: 0, paddingRight: 0, width: 0, height: 0)
+        
+        button.addSubview(rowCountLabel)
+        rowCountLabel.anchor(top: label.topAnchor, left: nil, bottom: label.bottomAnchor, right: button.rightAnchor, paddingTop: 0, paddingLeft: 0, paddingBottom: 0, paddingRight: 8, width: 0, height: 0)
         
         return view
     }
@@ -135,59 +157,38 @@ extension ObjectivesController {
     }
     
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: cellId, for: indexPath) as! ObjectiveCell
+        var cell = UITableViewCell()
         
-        let objective = allObjectives[indexPath.section][indexPath.row]
+        let objective = allObjectives[indexPath.section].objectives[indexPath.row]
 
 //        let objective = objectives[indexPath.row]
-        cell.objective = objective
+//        cell.objective = objective
         
         let backgroundView = UIView()
         
         if indexPath.section == 0 {
-            backgroundView.backgroundColor = UIColor.mainBlue
+            let dailyCell = tableView.dequeueReusableCell(withIdentifier: dailyCellId, for: indexPath) as! DailyCell
+            dailyCell.objective = objective
             
-            cell.unitLabel.removeFromSuperview()
-            cell.currentProgress.removeFromSuperview()
-            
-            cell.addSubview(cell.slider)
-            cell.slider.anchor(top: nil, left: cell.objectiveLabel.rightAnchor, bottom: nil, right: cell.rightAnchor, paddingTop: 0, paddingLeft: 4, paddingBottom: 0, paddingRight: 8, width: 0, height: 0)
-            cell.slider.centerYAnchor.constraint(equalTo: cell.centerYAnchor).isActive = true
-            
+            cell = dailyCell
         } else if indexPath.section == 1 {
+            let weeklyCell = tableView.dequeueReusableCell(withIdentifier: cellId, for: indexPath) as! ObjectiveCell
+            weeklyCell.objective = objective
             backgroundView.backgroundColor = UIColor.mainLightGreen
             
-            cell.slider.removeFromSuperview()
-            
-            cell.addSubview(cell.unitLabel)
-            cell.unitLabel.anchor(top: cell.objectiveLabel.topAnchor, left: nil, bottom: cell.objectiveLabel.bottomAnchor, right: cell.rightAnchor, paddingTop: 0, paddingLeft: 0, paddingBottom: 0, paddingRight: 4, width: 50, height: 0)
-
-            cell.addSubview(cell.currentProgress)
-            cell.currentProgress.anchor(top: nil, left: nil, bottom: nil, right: cell.unitLabel.leftAnchor, paddingTop: 0, paddingLeft: 0, paddingBottom: 0, paddingRight: 4, width: 0, height: 0)
-            cell.currentProgress.centerYAnchor.constraint(equalTo: cell.centerYAnchor).isActive = true
-            
+            cell = weeklyCell
         } else if indexPath.section == 2 {
+            let monthlyCell = tableView.dequeueReusableCell(withIdentifier: cellId, for: indexPath) as! ObjectiveCell
+            monthlyCell.objective = objective
             backgroundView.backgroundColor = UIColor.mainLightOrange
             
-            cell.slider.removeFromSuperview()
-            
-            cell.addSubview(cell.unitLabel)
-            cell.unitLabel.anchor(top: cell.objectiveLabel.topAnchor, left: nil, bottom: cell.objectiveLabel.bottomAnchor, right: cell.rightAnchor, paddingTop: 0, paddingLeft: 0, paddingBottom: 0, paddingRight: 4, width: 50, height: 0)
-
-            cell.addSubview(cell.currentProgress)
-            cell.currentProgress.anchor(top: nil, left: nil, bottom: nil, right: cell.unitLabel.leftAnchor, paddingTop: 0, paddingLeft: 0, paddingBottom: 0, paddingRight: 4, width: 0, height: 0)
-            cell.currentProgress.centerYAnchor.constraint(equalTo: cell.centerYAnchor).isActive = true
+            cell = monthlyCell
         } else if indexPath.section == 3 {
-            backgroundView.backgroundColor = UIColor.mainRed
+            let yearlyCell = tableView.dequeueReusableCell(withIdentifier: cellId, for: indexPath) as! ObjectiveCell
+            yearlyCell.objective = objective
+            backgroundView.backgroundColor = UIColor.mainLightOrange
             
-            cell.slider.removeFromSuperview()
-            
-//            cell.addSubview(cell.unitLabel)
-//            cell.unitLabel.anchor(top: cell.objectiveLabel.topAnchor, left: nil, bottom: cell.objectiveLabel.bottomAnchor, right: cell.rightAnchor, paddingTop: 0, paddingLeft: 0, paddingBottom: 0, paddingRight: 4, width: 50, height: 0)
-//
-//            cell.addSubview(cell.currentProgress)
-//            cell.currentProgress.anchor(top: nil, left: nil, bottom: nil, right: cell.unitLabel.leftAnchor, paddingTop: 0, paddingLeft: 0, paddingBottom: 0, paddingRight: 4, width: 0, height: 0)
-//            cell.currentProgress.centerYAnchor.constraint(equalTo: cell.centerYAnchor).isActive = true
+            cell = yearlyCell
         }
         
         cell.selectedBackgroundView = backgroundView
@@ -196,7 +197,10 @@ extension ObjectivesController {
     }
     
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return allObjectives[section].count
+        if !allObjectives[section].isExpanded {
+            return 0
+        }
+        return allObjectives[section].objectives.count
     }
     
     override func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
